Create thesisTemplate.tex
So this is a version of the master tex file I used for my thesis, so all the chapters/sections were in seperate files etc. It formats the spacing (Thank you Ed Butler!), and lets you play with the title page. It also has the authorship page and the abstract as described in the handbook and has the roman numerals in the appropriate place in the contents. The list of symbols is a bit tedious (Sorry...).

Hope this helps!